semantic-workflow/requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# AI/ML
openai>=1.0.0
google-cloud-language>=2.11.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Optional: Wikipedia enrichment
requests>=2.31.0
beautifulsoup4>=4.12.0

# Optional: PDF export
reportlab>=4.0.0
